+++ packages/blade/src/components/DatePicker/BaseDatePicker.web.tsx
@@ -82,6 +82,7 @@ const BaseDatePicker = <Type extends DateSelectionType = 'single'>({
   format = 'DD/MM/YYYY',
   inputPlaceHolder,
   inputElementType = 'datePickerInput',
+  showClearButton,
   onClearButtonClick,
   labelSuffix,
   labelTrailing,
@@ -93,7 +94,6 @@ const BaseDatePicker = <Type extends DateSelectionType = 'single'>({
   StyledPropsBlade &
   DataAnalyticsAttribute & {
     inputElementType: 'chip' | 'datePickerInput';
-    onClearButtonClick?: () => void;
   }): React.ReactElement => {
   const { i18nState } = useI18nContext();
   const _selectionType = selectionType ?? 'single';
@@ -264,6 +264,7 @@ const BaseDatePicker = <Type extends DateSelectionType = 'single'>({
   const handleClear = (): void => {
     fireNativeEvent(referenceRef, ['change']);
     handleReset();
+    setSelectedPreset(null);
     close();
     setFilterChipGroupSelectedFilters((prev: string[]) =>
       prev.filter((filter) => filter !== label),
@@ -505,12 +506,14 @@ const BaseDatePicker = <Type extends DateSelectionType = 'single'>({
                 excludeDate={props.excludeDate}
                 minDate={props.minDate}
                 maxDate={props.maxDate}
+                showClearButton={showClearButton}
+                onClearButtonClick={handleClear}
                 // Effective Selection type should only be use for selectionType 'range'
                 effectiveSelectionType={isSingle ? selectionType : effectiveSelectionType}
                 // Pass through preset state for showing label instead of date
                 selectedPresetLabel={selectedPresetLabel}
                 leadingDropdown={
-                  presets && !isSingle
+                  presets && !isSingle && hasBothDatesSelected
                     ? renderPresetDropdown({
                         onSelection: (preset: (date: Date) => DatesRangeValue) => {
                           const presetValue = preset?.(currentDate);